Req 08 fixed

+++ desafio8.sql
@@ -1,14 +1,10 @@
-USE SpotifyClone;
-
 DELIMITER $$
-CREATE 
-    TRIGGER  trigger_usuario_delete
-AFTER DELETE ON Users FOR EACH ROW 
+
+CREATE TRIGGER trigger_usuario_delete
+BEFORE DELETE ON Users FOR EACH ROW 
     BEGIN
-      DELETE FROM Followers WHERE Followers.user_id = OLD.Followers.user_id;
-DELETE FROM Playback_History 
-WHERE
-    Playback_History.user_id = OLD.Playback_History.user_id;
+      DELETE FROM Followers WHERE user_id = OLD.user_id;
+      DELETE FROM Playback_History WHERE user_id = OLD.user_id;
     END$$
 
 DELIMITER ;
